requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-cosmos==4.9.0
azure-storage-blob==12.25.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
isodate==0.7.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
